src/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
kombu==5.3.5
numpy==1.26.2
Pillow==10.1.0
pillow_heif==0.15.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.13
psycopg2-binary
gunicorn
opencv-python-headless